Bugfix regarding casing when transpositions are enabled
- When a transposition was set, and an ascii text was attempted to be
  encrypted, if the ascii was lowercase letters, it wouldn't be
transformed to upper before being mapped through the transpositions
but only after it went through the rotors.

This caused a situation where the encrypted result would be wrong since
every lower case character never passed through the transpositions but
emerged as upper case. When deciphered, the result would be complete
garbage as now the input is already upper case and everything will pass
through the transpositions.